fix: use int64 for the chunk index (#746)
While it's not a concern for one NAR to be chunked into more chunks than
an int32 can handle, the performance and storage gain is simply not
worth the special handling I had to do in order to support int32.
Convert to int64 to keep the code simpler and the database essentially
limitless in terms of number of chunks.

+++ db/migrations/mysql/20260131021850_add_chunks.sql
@@ -10,7 +10,7 @@ CREATE TABLE chunks (
 CREATE TABLE nar_file_chunks (
     nar_file_id BIGINT NOT NULL,
     chunk_id BIGINT NOT NULL,
-    chunk_index INT NOT NULL,
+    chunk_index BIGINT NOT NULL,
     PRIMARY KEY (nar_file_id, chunk_index),
     FOREIGN KEY (nar_file_id) REFERENCES nar_files (id) ON DELETE CASCADE,
     FOREIGN KEY (chunk_id) REFERENCES chunks (id) ON DELETE CASCADE
