+++ db/migrations/postgres/20260205063658_add_total_chunks_to_nar_files.sql
@@ -1,5 +1,5 @@
 -- migrate:up
-ALTER TABLE nar_files ADD COLUMN total_chunks INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE nar_files ADD COLUMN total_chunks BIGINT NOT NULL DEFAULT 0;
 
 -- migrate:down
 ALTER TABLE nar_files DROP COLUMN total_chunks;
