+++ pkg/database/postgresdb/models.go
@@ -34,13 +34,13 @@ type NarFile struct {
 	CreatedAt      time.Time
 	UpdatedAt      sql.NullTime
 	LastAccessedAt sql.NullTime
-	TotalChunks    int32
+	TotalChunks    int64
 }
 
 type NarFileChunk struct {
 	NarFileID  int64
 	ChunkID    int64
-	ChunkIndex int32
+	ChunkIndex int64
 }
 
 type NarInfo struct {
